Time: 8 ms (55.26%), Space: 13.9 MB (17.29%) - LeetHub

+++ 152-maximum-product-subarray/152-maximum-product-subarray.cpp
@@ -1,15 +1,15 @@
 class Solution {
 public:
     int maxProduct(vector<int>& nums) {
-        long large = nums.front();
-        long small = nums.front();
-        long result = large;
+        int large = nums.front();
+        int small = nums.front();
+        int result = large;
         for(int i = 1; i < nums.size();i++){
-            long small_take_new = small * nums[i];
-            long large_take_new = large * nums[i];
+            int small_take_new = small * nums[i];
+            int large_take_new = large * nums[i];
             
-            small = min(min((long)nums[i], small_take_new),large_take_new);
-            large = max(max((long)nums[i], large_take_new),small_take_new);
+            small = min(min(nums[i], small_take_new),large_take_new);
+            large = max(max(nums[i], large_take_new),small_take_new);
             
             result = max(large,result);
         }
